Initial template transformation
This uses handlebars to substitute for job variables provided by
gitlab, with the prefix of `job.`, so to substitute for
`CI_JOB_TOKEN` you can use `{{ job.CI_JOB_TOKEN }}`.

Note that this is the runner, and template transformations may already
have been applied at an earlier generation step for the job
(e.g. using jinja) which means the actual substitutions in this phase
may need to have been escaped.