Throttle the rate of LAVA API requests
We need to avoid saturating a Lava instance with more requests than it
can handle when we have many jobs to monitor. To do so, we limit the
number of concurrent API endpoints that are active to 20. This could
easily be made configurable in future.

To achieve this, we wrap the Lava instance in a ThrottledLava shell
that requires a permit to allow each API call to be made. The permits
are passed on to the return values and need to be held until the API
call finishes.

The permits themselves are granted by a Throttler, which has an
underlying semaphore with limited capacity. Many endpoints are handled
with a simple Throttled wrapper type, which is just a smart pointer
that holds a Permit. The types that cannot be handled this way are the
builders.