Add debug logging
These help significantly in debugging the late release of tokens which
can cause querying to stall as the request volume grows.

A previous iteration of the code held the Jobs endpoint open
throughout its main loop, because it performed its work within a loop
that only has one iteration. During this loop, the logging endpoint
was called, which is also rate limited.

Because there can be many jobs active, sharing the same rate limiting,
it's possible for all the entire system to deadlock on the
inaccessibility of permits, and in fact it appears to be a
precondition for successful operation of this system that nothing ever
tries to obtain two permits simultaneously.

The precondition is difficult to check or track, but fortunately
debugging it is relatively straightforward when the
--max-concurrent-jobs is set low (which is a more stringent situation
that places more demand on permits, in the presence of the same number
of clients). This is the rationale for including this code, despite
its invasive nature.