Make the logs more readable
They were being printed for consumption in Gitlab using just the debug
display for each entry. That makes it really hard to read, and
especially to identify failures.

This adds some actual formatting for messages to remove some of the
noise, and puts the results in a box to make them easier to spot. This
is essentially an ASCII version of the display LAVA itself gives.