Support for Gitlab cancellation
At present, if a job is cancelled in Gitlab, the LAVA jobs are not
cancelled; in fact the framework makes no attempt to detect
cancellation or respond to it. This can leave LAVA submission queues
clogged with jobs whose output is not of interest.

The cancellation behaviour can be customised by setting
LAVA_CANCELLATION in the Gitlab job. The options are 'cancel' to
cancel any pending LAVA jobs, and 'ignore' to leave the LAVA jobs
running. The default behaviour is to cancel jobs run via 'submit', and
to leave jobs running for jobs running under 'monitor'.

Signed-off-by: Ed Smith <ed.smith@collabora.com>