Use the new artifact model in gitlab-runner-rs
gitlab-runner-rs now has a model in which it requests a list of files
that we could upload, and then performs globbing (as the upstream
runner does) to determine which the user has selected, and finally
groups these files together into artifacts.

By moving to this updated model, we enable our junit test results to
be properly integrated with Gitlab, and show up in the pipeline view.
There's no direct code support required for this in this crate, we
just have to switch to the more sophisticated model that the
gitlab-runner crate has introduced.