Retrieve Junit format test results from LAVA
This requires an update to the lava-api crate to be able to retrieve
the test results.

There are two additional uses for these logs:

- We can use the test results to correct the status report we provide
for test runs. Currently, we return 'success' if the test completes,
regardless of whether any tests fail. By parsing the junit report,
we can switch to reporting success only when no tests fail.

- We can upload junit reports as artifacts to Gitlab. The support to
integrate these reports properly into Gitlab will be added in a
subsequent commit. This just adds any test reports as files to the
uploaded artifact archive.