Switch to the authentication token-based runner creation workflow
Starting in GitLab v15.10 (and enabled by default in v16.0), the runner
creation uses a workflow based on authentication tokens instead of
registration tokens. This introduces a new runner property named
system_id that needs to be passed to the /jobs/request API.

Extend the runner with a new system_id parameter (passed through the
environment variable 'RUNNER_SYSTEM_ID'. Additionall update the
documentation in README.md to explain the new workflow.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>