ci: switch away from actions-rs
actions-rs is no longer maintained, switch over to dtolnay/rust-toolchain
instead which is. While there also ensure the various checks also check
all targets/features as applicable

For the rust toolchain aim for a specific version so CI doesn't start
failing over time. For the version also aim for the version of rustc in
debian unstable to ensure things can build with that version of the
compiler